refactor(user): split user type tests for GH727
